move config generation to config.go
